move save by location init into std late game manager init
