+++ SolastaCommunityExpansion/Models/SaveByLocationContext.cs
@@ -1,4 +1,5 @@
-﻿using System.IO;
+﻿using System;
+using System.IO;
 using System.Linq;
 
 namespace SolastaCommunityExpansion.Models
@@ -93,7 +94,70 @@ namespace SolastaCommunityExpansion.Models
             }
         }
 
-        public static int SaveFileCount(LocationType locationType, string folder)
+        internal static void Load()
+        {
+            if (!Main.Settings.EnableSaveByLocation)
+            {
+                return;
+            }
+
+            // Ensure folders exist
+            Directory.CreateDirectory(LocationSaveGameDirectory);
+            Directory.CreateDirectory(CampaignSaveGameDirectory);
+
+            // Find the most recently touched save file and select the correct location/campaign for that save
+            var mostRecent = Directory.EnumerateDirectories(LocationSaveGameDirectory)
+                .Select(d => new
+                {
+                    Path = d,
+                    LastWriteTime = Directory.EnumerateFiles(d, "*.sav").Max(f => (DateTime?)File.GetLastWriteTimeUtc(f)),
+                    LocationType = LocationType.UserLocation
+                })
+                .Concat(
+                    Directory.EnumerateDirectories(CampaignSaveGameDirectory)
+                        .Select(d => new
+                        {
+                            Path = d,
+                            LastWriteTime = Directory.EnumerateFiles(d, "*.sav").Max(f => (DateTime?)File.GetLastWriteTimeUtc(f)),
+                            LocationType = LocationType.CustomCampaign
+                        })
+                )
+                .Concat(
+                    Enumerable.Repeat(
+                        new
+                        {
+                            Path = DefaultSaveGameDirectory,
+                            LastWriteTime = Directory.EnumerateFiles(DefaultSaveGameDirectory, "*.sav").Max(f => (DateTime?)File.GetLastWriteTimeUtc(f)),
+                            LocationType = LocationType.StandardCampaign
+                        }
+                    , 1)
+                )
+                .Where(d => d.LastWriteTime.HasValue)
+                .OrderByDescending(d => d.LastWriteTime)
+                .FirstOrDefault();
+
+            var selectedCampaignService = ServiceRepositoryEx.GetOrCreateService<SelectedCampaignService>();
+
+            if (selectedCampaignService != null && mostRecent != null)
+            {
+                Main.Log($"Most recent folder={mostRecent.Path}");
+
+                switch (mostRecent.LocationType)
+                {
+                    default:
+                        selectedCampaignService.SetStandardCampaignLocation();
+                        break;
+                    case LocationType.UserLocation:
+                        selectedCampaignService.SetCampaignLocation(USER_CAMPAIGN, Path.GetFileName(mostRecent.Path));
+                        break;
+                    case LocationType.CustomCampaign:
+                        selectedCampaignService.SetCampaignLocation(Path.GetFileName(mostRecent.Path), string.Empty);
+                        break;
+                }
+            }
+        }
+
+        internal static int SaveFileCount(LocationType locationType, string folder)
         {
             switch (locationType)
             {
